+++ PSP/src/com/porcupine/psp/model/service/ImplSeguridadService.java
@@ -42,15 +42,15 @@ public class ImplSeguridadService implements IService<ImplSeguridadVO, Short> {
         entity.setDescripcionI(vo.getDescripcionI());
         entity.setEstadoI(vo.getEstadoI());
         entity.setFechaRegIm(vo.getFechaRegIm());
-        
-//        Proveedor proveedor = DAOFactory.getInstance().getProveedorDAO().find(new Integer(vo.getIdPro()));
+       
+//        Proveedor proveedor = DAOFactory.getInstance().getProveedorDAO().find(vo.getIdPro());
 //        entity.setIdPro(proveedor);
 //        
 //        CoordTYT coordinadorTyT = DAOFactory.getInstance().getCoordTYTDAO().find(vo.getCedulaCoordTyT());
 //        entity.setCedulae(coordinadorTyT);
 
         if (vo.getIdPro() != null) {
-            Proveedor proveedor = DAOFactory.getInstance().getProveedorDAO().find((int) vo.getIdPro());
+            Proveedor proveedor = DAOFactory.getInstance().getProveedorDAO().find(vo.getIdPro());
             entity.setIdPro(proveedor);
             proveedor.getImplSeguridadList().add(entity);
         }
@@ -85,7 +85,7 @@ public class ImplSeguridadService implements IService<ImplSeguridadVO, Short> {
         entity.setFechaRegIm(vo.getFechaRegIm());
 
         if (vo.getIdPro() != 0) {
-            Proveedor proveedor = DAOFactory.getInstance().getProveedorDAO().find((int) vo.getIdPro());
+            Proveedor proveedor = DAOFactory.getInstance().getProveedorDAO().find(vo.getIdPro());
             entity.setIdPro(proveedor);
             proveedor.getImplSeguridadList().set(proveedor.getImplSeguridadList().get(entity.getIdImplemento()).getIdImplemento(), entity);
         }
